requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - CORRIGIDO
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.10
alembic==1.13.1

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# Redis
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.25.2

# Validation
email-validator==2.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Utils
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pytest==8.4.1
pytest-tornasync==0.6.0.post2